third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/object-embed-tag-expected.txt

CONSOLE MESSAGE: line 9: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?relay-target-ids-for-event=beforeload&q=%3Cobject%20id=%27object%27%20name=%27plugin%27%20type=%27application/x-webkit-test-netscape%27%3E%3Cparam%20name=%27movie%27%20value=%27http://127.0.0.1:8000/security/xssAuditor/resources/dummy.swf%27%20/%3E%3Cembed%20id=%27embed%27%20name=%27plugin%27%20type=%27application/x-webkit-test-netscape%27%20src=%27http://127.0.0.1:8000/security/xssAuditor/resources/dummy.swf%27%20/%3E%3C/object%3E' because its source code was found within the request. The auditor was enabled as the server sent neither an 'X-XSS-Protection' nor 'Content-Security-Policy' header.
CONSOLE MESSAGE: line 9: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?relay-target-ids-for-event=beforeload&q=%3Cobject%20id=%27object%27%20name=%27plugin%27%20type=%27application/x-webkit-test-netscape%27%3E%3Cparam%20name=%27movie%27%20value=%27http://127.0.0.1:8000/security/xssAuditor/resources/dummy.swf%27%20/%3E%3Cembed%20id=%27embed%27%20name=%27plugin%27%20type=%27application/x-webkit-test-netscape%27%20src=%27http://127.0.0.1:8000/security/xssAuditor/resources/dummy.swf%27%20/%3E%3C/object%3E' because its source code was found within the request. The auditor was enabled as the server sent neither an 'X-XSS-Protection' nor 'Content-Security-Policy' header.
CONSOLE MESSAGE: line 9: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?relay-target-ids-for-event=beforeload&q=%3Cobject%20id=%27object%27%20name=%27plugin%27%20type=%27application/x-webkit-test-netscape%27%3E%3Cparam%20name=%27movie%27%20value=%27http://127.0.0.1:8000/security/xssAuditor/resources/dummy.swf%27%20/%3E%3Cembed%20id=%27embed%27%20name=%27plugin%27%20type=%27application/x-webkit-test-netscape%27%20src=%27http://127.0.0.1:8000/security/xssAuditor/resources/dummy.swf%27%20/%3E%3C/object%3E' because its source code was found within the request. The auditor was enabled as the server sent neither an 'X-XSS-Protection' nor 'Content-Security-Policy' header.
